rename logger function to log
